requirements.txt
gradio
transformers
torch
Pillow
numpy